+++ resolwe/flow/serializers/entity.py
@@ -4,20 +4,15 @@ from django.db import transaction
 from resolwe.flow.models import Collection, Entity
 from resolwe.permissions.models import Permission
 
-from .collection import BaseCollectionSerializer
+from .collection import BaseCollectionSerializer, CollectionSerializer
 from .fields import DictRelatedField
 
 
-class EntitySerializer(BaseCollectionSerializer):
-    """Serializer for Entity."""
+class BaseEntitySerializer(BaseCollectionSerializer):
+    """Entity serializer without the collection field.
 
-    collection = DictRelatedField(
-        queryset=Collection.objects.all(),
-        serializer=BaseCollectionSerializer,
-        allow_null=True,
-        required=False,
-        write_permission=Permission.EDIT,
-    )
+    Used when serializing data object, since collection info is duplicated.
+    """
 
     class Meta(BaseCollectionSerializer.Meta):
         """EntitySerializer Meta options."""
@@ -38,3 +33,15 @@ class EntitySerializer(BaseCollectionSerializer):
         if update_collection and new_collection != instance.collection:
             instance.move_to_collection(new_collection)
         return instance
+
+
+class EntitySerializer(BaseEntitySerializer):
+    """Serializer for Entity with added collection field."""
+
+    collection = DictRelatedField(
+        queryset=Collection.objects.all(),
+        serializer=CollectionSerializer,
+        allow_null=True,
+        required=False,
+        write_permission=Permission.EDIT,
+    )